Project_Recipe_generator_v1
This is the first modification of the base code.

It asks many questions but I am still not able to include the extra filters in the query to the edamam database.
I have added "I" to have the number of recipes. Hopefully will be useful to select the file we want to export.